Remove some unnecessary request fields
The purpose of "tags" was to distinguish between planned and unplanned
activities. We can still do that if we like, but I'd structure it
differently.

The purpose of "parallel_activities" was to be able to set activities
which must happen concurrently. It's my understanding that this
doesn't apply to our use case.